% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fct_adjust_get_years.R
\name{fct_adjust_get_years}
\alias{fct_adjust_get_years}
\title{Adjust Max Year Based on Current Year and Day of Year}
\usage{
fct_adjust_get_years(years)
}
\arguments{
\item{years}{list of year ranges sourced from get_years() reactive}

\item{max_year}{maximum outmigration year of the input data}
}
\value{
range of years
}
\description{
returns same range from reactive get_years() based on speciesxrun but adjusts max year to remove years where all adults have not returned (i.e. 3 years past last outmigration year)
}
