% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydroSurvDOYTEMP-package.R
\docType{package}
\name{HydroSurvDOYTEMP-package}
\alias{HydroSurvDOYTEMP}
\alias{HydroSurvDOYTEMP-package}
\title{HydroSurvDOYTEMP: Seasonal Predictions of SAR & T:I for Chinook salmon and Steelhead}
\description{
PKG_DESC.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/caitobrien/HydroSurvDOYTEMP}
  \item Report bugs at \url{https://github.com/caitobrien/HydroSurvDOYTEMP/issues}
}

}
\author{
\strong{Maintainer}: CAITLIN OBRIEN \email{CSOBRIEN@UW.EDU}

}
\keyword{internal}
